+++ app/mailers/emails/groups.rb
@@ -1,0 +1,11 @@
+module Emails
+  module Groups
+    def group_access_granted_email(user_group_id)
+      @membership = UsersGroup.find(user_group_id)
+      @group = @membership.group
+
+      mail(to: @membership.user.email,
+           subject: subject("access to group was granted"))
+    end
+  end
+end
