+++ app/views/notify/group_access_granted_email.html.haml
@@ -1,0 +1,5 @@
+%p
+  = "You have been granted #{@membership.human_access} access to group"
+%p
+  = link_to group_url(@group) do
+    = @group.name
