+++ app/views/notify/group_access_granted_email.text.erb
@@ -1,0 +1,4 @@
+
+You have been granted <%= @membership.human_access %> access to group <%= @group.name %>
+
+<%= url_for(group_url(@group)) %>
